+++ solutions/src/p01_ring_buffer.c
@@ -1,11 +1,10 @@
 #include "p01_ring_buffer.h"
-#include <stdlib.h>
 /* TODO: implement. Remember: cap must be power of two.
  * index into buf = value & (cap - 1)   <-- no division needed */
-int rb_init(RingBuffer *rb, uint32_t cap) { return -1; }
+int rb_init(RingBuffer *rb, size_t cap) { return -1; }
 void rb_destroy(RingBuffer *rb) {}
 int rb_push(RingBuffer *rb, uint8_t byte) { return -1; }
 int rb_pop(RingBuffer *rb, uint8_t *out) { return -1; }
-uint32_t rb_used(const RingBuffer *rb) { return 0; }
+size_t rb_used(const RingBuffer *rb) { return 0; }
 bool rb_is_empty(const RingBuffer *rb) { return true; }
 bool rb_is_full(const RingBuffer *rb) { return false; }
